# docs/api/backend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopSteward Backend API — 设计契约（尚未实现）",
    "version": "0.2.0-draft",
    "description": "B0独立业务后端；B2为未来接口。Swagger展示不代表服务已实现。业务语义见../backend-development.md。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "规划的本地backend地址，当前未启动"
    }
  ],
...
  "paths": {
...
    "/api/v1/missions": {
      "post": {
        "operationId": "create_mission",
        "tags": [
          "Missions"
        ],
        "summary": "创建任务并登记首次检查",
        "description": "创建任务并登记首次检查",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "operator"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                },
                "example": {
                  "id": "mission_001",
                  "store_id": "store_001",
                  "sku_id": "sku_001",
                  "objective": "在现金底线约束下覆盖活动需求",
                  "status": "ACTIVE",
                  "mission_version": 1,
                  "policy": {
                    "cash_floor_minor": 30000,
                    "candidate_quantities": [
                      0,
                      20,
                      40,
                      80
                    ],
                    "supplier_id": "supplier_001"
                  },
                  "policy_version": "policy-1",
                  "schedule": {
                    "id": "schedule_001",
                    "job_type": "check_mission",
                    "mission_id": "mission_001",
                    "trigger": "INTERVAL",
                    "interval_seconds": 30,
                    "enabled": true,
                    "version": 1,
                    "next_run_at": "2026-09-06T00:00:30Z"
                  },
                  "current_plan_id": "plan_001",
                  "current_action_id": null,
                  "completion_criteria": "用户请求结束且无未决采购动作",
                  "created_at": "2026-09-06T00:00:00Z",
                  "updated_at": "2026-09-06T00:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              },
              "example": {
                "store_id": "store_001",
                "sku_id": "sku_001",
                "objective": "现金不少于300元并覆盖活动需求",
                "policy": {
                  "cash_floor_minor": 30000,
                  "candidate_quantities": [
                    0,
                    20,
                    40,
                    80
                  ],
                  "supplier_id": "supplier_001"
                },
                "check_interval_seconds": 30
              }
            }
          }
        }
      },
      "get": {
        "operationId": "list_missions",
        "tags": [
          "Missions"
        ],
        "summary": "分页查询任务",
        "description": "分页查询任务",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PAUSED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionList"
                },
                "example": {
                  "items": [
                    {
                      "id": "mission_001",
                      "store_id": "store_001",
                      "sku_id": "sku_001",
                      "objective": "在现金底线约束下覆盖活动需求",
                      "status": "ACTIVE",
                      "mission_version": 1,
                      "policy": {
                        "cash_floor_minor": 30000,
                        "candidate_quantities": [
                          0,
                          20,
                          40,
                          80
                        ],
                        "supplier_id": "supplier_001"
                      },
                      "policy_version": "policy-1",
                      "schedule": {
                        "id": "schedule_001",
                        "job_type": "check_mission",
                        "mission_id": "mission_001",
                        "trigger": "INTERVAL",
                        "interval_seconds": 30,
                        "enabled": true,
                        "version": 1,
                        "next_run_at": "2026-09-06T00:00:30Z"
                      },
                      "current_plan_id": "plan_001",
                      "current_action_id": null,
                      "completion_criteria": "用户请求结束且无未决采购动作",
                      "created_at": "2026-09-06T00:00:00Z",
                      "updated_at": "2026-09-06T00:00:00Z"
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}": {
      "get": {
        "operationId": "get_mission",
        "tags": [
          "Missions"
        ],
        "summary": "查询任务详情",
        "description": "查询任务详情",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer",
          "agent:read"
        ],
        "security": [
          {
            "UserBearer": []
          },
          {
            "ServiceBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                },
                "example": {
                  "id": "mission_001",
                  "store_id": "store_001",
                  "sku_id": "sku_001",
                  "objective": "在现金底线约束下覆盖活动需求",
                  "status": "ACTIVE",
                  "mission_version": 1,
                  "policy": {
                    "cash_floor_minor": 30000,
                    "candidate_quantities": [
                      0,
                      20,
                      40,
                      80
                    ],
                    "supplier_id": "supplier_001"
                  },
                  "policy_version": "policy-1",
                  "schedule": {
                    "id": "schedule_001",
                    "job_type": "check_mission",
                    "mission_id": "mission_001",
                    "trigger": "INTERVAL",
                    "interval_seconds": 30,
                    "enabled": true,
                    "version": 1,
                    "next_run_at": "2026-09-06T00:00:30Z"
                  },
                  "current_plan_id": "plan_001",
                  "current_action_id": null,
                  "completion_criteria": "用户请求结束且无未决采购动作",
                  "created_at": "2026-09-06T00:00:00Z",
                  "updated_at": "2026-09-06T00:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/control": {
      "post": {
        "operationId": "control_mission",
        "tags": [
          "Missions"
        ],
        "summary": "暂停、恢复、完成或取消任务",
        "description": "暂停、恢复、完成或取消任务",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "operator:pause_resume",
          "approver:complete_cancel"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                },
                "example": {
                  "id": "mission_001",
                  "store_id": "store_001",
                  "sku_id": "sku_001",
                  "objective": "在现金底线约束下覆盖活动需求",
                  "status": "ACTIVE",
                  "mission_version": 1,
                  "policy": {
                    "cash_floor_minor": 30000,
                    "candidate_quantities": [
                      0,
                      20,
                      40,
                      80
                    ],
                    "supplier_id": "supplier_001"
                  },
                  "policy_version": "policy-1",
                  "schedule": {
                    "id": "schedule_001",
                    "job_type": "check_mission",
                    "mission_id": "mission_001",
                    "trigger": "INTERVAL",
                    "interval_seconds": 30,
                    "enabled": true,
                    "version": 1,
                    "next_run_at": "2026-09-06T00:00:30Z"
                  },
                  "current_plan_id": "plan_001",
                  "current_action_id": null,
                  "completion_criteria": "用户请求结束且无未决采购动作",
                  "created_at": "2026-09-06T00:00:00Z",
                  "updated_at": "2026-09-06T00:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionControl"
              },
              "example": {
                "operation": "pause",
                "expected_mission_version": 1
              }
            }
          }
        }
      }
    },
...
    "/api/v1/missions/{mission_id}/checks": {
      "post": {
        "operationId": "request_mission_check",
        "tags": [
          "Scheduling"
        ],
        "summary": "请求一次检查，可合并已有待执行检查",
        "description": "请求一次检查，可合并已有待执行检查",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "operator",
          "agent:check"
        ],
        "security": [
          {
            "UserBearer": []
          },
          {
            "ServiceBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "同一逻辑命令网络重试复用；同键不同内容返回409。",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "202": {
            "description": "已持久化受理，不表示业务执行完成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                },
                "example": {
                  "job_run_id": "job_001",
                  "status": "READY",
                  "merged": false
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本、幂等或业务并发冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              },
              "example": {
                "reason": "用户要求再次检查"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "用户身份；角色与门店范围由服务端验证。"
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "服务身份；Agent无审批权限，simulation仅提交授权事件。"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "MissionCreate": {
        "type": "object",
        "properties": {
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sku_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "objective": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "check_interval_seconds": {
            "type": "integer",
            "minimum": 5,
            "maximum": 3600,
            "default": 30
          }
        },
        "required": [
          "store_id",
          "sku_id",
          "objective",
          "policy",
          "check_interval_seconds"
        ],
        "additionalProperties": false
      },
      "MissionControl": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "pause",
              "resume",
              "complete",
              "cancel"
            ]
          },
          "expected_mission_version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "operation",
          "expected_mission_version"
        ],
        "additionalProperties": false
      },
...
      "Mission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "sku_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "objective": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "mission_version": {
            "type": "integer",
            "minimum": 1
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "policy_version": {
            "type": "string"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "current_plan_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "current_action_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "completion_criteria": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "store_id",
          "sku_id",
          "objective",
          "status",
          "mission_version",
          "policy",
          "policy_version",
          "schedule",
          "current_plan_id",
          "current_action_id",
          "completion_criteria",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "CheckRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "用户要求重新检查的原因",
            "maxLength": 1000
          }
        },
        "required": [],
        "additionalProperties": false
      },
      "JobAccepted": {
        "type": "object",
        "properties": {
          "job_run_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "RUNNING",
              "RETRY_WAIT"
            ]
          },
          "merged": {
            "type": "boolean"
          }
        },
        "required": [
          "job_run_id",
          "status",
          "merged"
        ],
        "additionalProperties": false
      },
...
      "MissionList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mission"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}